requirements.txt
PyPDF2
opnenai
python-dotenv
